requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
pinecone-client>=3.0.0
openai>=1.0.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
